New update I'm pushing: I created an admin account which will be used for login testing. This admin account I anticipate will stay even till the end.
NEW CODE:
NewUserTest
This will test the creation of a new user, and the detection of a duplicate user entry.
This will also do away with having to manually enter a new user id
It will take the previous user id (including manager id if the user is a manager) and increase it by 1 and assign this id to the new user.
It works with gradle testing.
I have not yet implemented the more efficient query method Ali has referred to the other day. I will do this probably later tonight or tomorrow. I just wanted to get this code working the way I knew how to first.
I am pushing this update now